requirements.txt
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
khmer-nltk==1.6
nltk==3.9.1
numpy==2.2.6
python-crfsuite==0.9.11
python-dotenv==1.1.0
python-telegram-bot==22.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.3
sklearn-crfsuite==0.5.0
sniffio==1.3.1
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
